elasticsearcl-client/src/main/java/com/zx/learn/es/client/dao/EsIndexDao.java
package com.zx.learn.es.client.dao;


import com.alibaba.fastjson.JSONObject;
import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
import org.elasticsearch.action.index.IndexRequest;
import org.elasticsearch.action.index.IndexResponse;
import org.elasticsearch.action.support.IndicesOptions;
import org.elasticsearch.action.support.master.AcknowledgedResponse;
import org.elasticsearch.client.RequestOptions;
import org.elasticsearch.client.RestHighLevelClient;
import org.elasticsearch.common.settings.Settings;
import org.elasticsearch.common.xcontent.XContentType;
import org.elasticsearch.rest.RestStatus;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import java.io.IOException;

/**
 * Description:
 *
 * @author: chixiao
 * @date: 2019-09-05
 * @time: 10:23
 */
@Component
public class EsIndexDao {

    @Autowired
    private RestHighLevelClient client;

    public boolean createIndex(String indexName, String source) throws IOException {
        CreateIndexRequest request = new CreateIndexRequest(indexName);
        request.source(source, XContentType.JSON);
        CreateIndexResponse response = client.indices().create(request, RequestOptions.DEFAULT);
        return response.isAcknowledged();
    }

    public boolean deleteIndex(String indexName) throws IOException {
        DeleteIndexRequest request = new DeleteIndexRequest(indexName);
        request.indicesOptions(IndicesOptions.lenientExpandOpen());
        AcknowledgedResponse response = client.indices().delete(request, RequestOptions.DEFAULT);
        return response.isAcknowledged();
    }

    public boolean addDocument(String indexName, String id, Object object) throws IOException {
        IndexRequest request = new IndexRequest(indexName, indexName);
        request.id(id);
        String json = JSONObject.toJSONString(object);
        request.source(json, XContentType.JSON);
        IndexResponse indexResponse = client.index(request, RequestOptions.DEFAULT);
        return indexResponse.status() == RestStatus.ACCEPTED;
    }

}
